% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centroidview.R
\name{centroidview}
\alias{centroidview}
\title{Study Centroids from a Hierarchical Clustering}
\usage{
centroidview(x, tree, groups, samples, width = NULL, height = NULL,
  elementId = NULL)
}
\description{
This creates an interactive heatmap + hierarchical clustering view.
}

